// api/TriviaGraphQL/generated.ts
export type Maybe<T> = T | null;
export type Exact<T extends { [key: string]: unknown }> = { [K in keyof T]: T[K] };
export type RequireFields<T, K extends keyof T> = { [X in Exclude<keyof T, K>]?: T[X] } & { [P in K]-?: NonNullable<T[P]> };
/** All built-in and custom scalars, mapped to their actual values */
export type Scalars = {
  ID: string;
  String: string;
  Boolean: boolean;
  Int: number;
  Float: number;
};

export type Question = {
  __typename?: 'Question';
  id: Scalars['ID'];
  question: Scalars['String'];
  correctAnswer: Scalars['String'];
  answers: Array<Scalars['String']>;
};

export type Query = {
  __typename?: 'Query';
  question?: Maybe<Question>;
  getRandomQuestion?: Maybe<Question>;
};


export type QueryQuestionArgs = {
  id: Scalars['ID'];
};

export type Answer = {
  __typename?: 'Answer';
  questionId?: Maybe<Scalars['ID']>;
  question: Scalars['String'];
  submittedAnswer: Scalars['String'];
  correctAnswer: Scalars['String'];
  correct?: Maybe<Scalars['Boolean']>;
};

export type Mutation = {
  __typename?: 'Mutation';
  answerQuestion?: Maybe<Answer>;
};


export type MutationAnswerQuestionArgs = {
  id?: Maybe<Scalars['ID']>;
  answer?: Maybe<Scalars['String']>;
};
// ...
export type Resolver<TResult, TParent = {}, TContext = {}, TArgs = {}> =
  | ResolverFn<TResult, TParent, TContext, TArgs>
  | StitchingResolver<TResult, TParent, TContext, TArgs>;
// ...
export type IsTypeOfResolverFn<T = {}, TContext = {}> = (obj: T, context: TContext, info: GraphQLResolveInfo) => boolean | Promise<boolean>;

export type NextResolverFn<T> = () => Promise<T>;

export type DirectiveResolverFn<TResult = {}, TParent = {}, TContext = {}, TArgs = {}> = (
  next: NextResolverFn<TResult>,
  parent: TParent,
  args: TArgs,
  context: TContext,
  info: GraphQLResolveInfo
) => TResult | Promise<TResult>;

/** Mapping between all available schema types and the resolvers types */
export type ResolversTypes = {
  Question: ResolverTypeWrapper<QuestionModel>;
  ID: ResolverTypeWrapper<Scalars['ID']>;
  String: ResolverTypeWrapper<Scalars['String']>;
  Query: ResolverTypeWrapper<{}>;
  Answer: ResolverTypeWrapper<Answer>;
  Boolean: ResolverTypeWrapper<Scalars['Boolean']>;
  Mutation: ResolverTypeWrapper<{}>;
};

/** Mapping between all available schema types and the resolvers parents */
export type ResolversParentTypes = {
  Question: QuestionModel;
  ID: Scalars['ID'];
  String: Scalars['String'];
  Query: {};
  Answer: Answer;
  Boolean: Scalars['Boolean'];
  Mutation: {};
};

export type QuestionResolvers<ContextType = Context, ParentType extends ResolversParentTypes['Question'] = ResolversParentTypes['Question']> = {
  id: Resolver<ResolversTypes['ID'], ParentType, ContextType>;
  question: Resolver<ResolversTypes['String'], ParentType, ContextType>;
  correctAnswer: Resolver<ResolversTypes['String'], ParentType, ContextType>;
  answers: Resolver<Array<ResolversTypes['String']>, ParentType, ContextType>;
  __isTypeOf?: IsTypeOfResolverFn<ParentType, ContextType>;
};

export type QueryResolvers<ContextType = Context, ParentType extends ResolversParentTypes['Query'] = ResolversParentTypes['Query']> = {
  question: Resolver<Maybe<ResolversTypes['Question']>, ParentType, ContextType, RequireFields<QueryQuestionArgs, 'id'>>;
  getRandomQuestion: Resolver<Maybe<ResolversTypes['Question']>, ParentType, ContextType>;
};

export type AnswerResolvers<ContextType = Context, ParentType extends ResolversParentTypes['Answer'] = ResolversParentTypes['Answer']> = {
  questionId: Resolver<Maybe<ResolversTypes['ID']>, ParentType, ContextType>;
  question: Resolver<ResolversTypes['String'], ParentType, ContextType>;
  submittedAnswer: Resolver<ResolversTypes['String'], ParentType, ContextType>;
  correctAnswer: Resolver<ResolversTypes['String'], ParentType, ContextType>;
  correct: Resolver<Maybe<ResolversTypes['Boolean']>, ParentType, ContextType>;
  __isTypeOf?: IsTypeOfResolverFn<ParentType, ContextType>;
};

export type MutationResolvers<ContextType = Context, ParentType extends ResolversParentTypes['Mutation'] = ResolversParentTypes['Mutation']> = {
  answerQuestion: Resolver<Maybe<ResolversTypes['Answer']>, ParentType, ContextType, RequireFields<MutationAnswerQuestionArgs, never>>;
};

export type Resolvers<ContextType = Context> = {
  Question: QuestionResolvers<ContextType>;
  Query: QueryResolvers<ContextType>;
  Answer: AnswerResolvers<ContextType>;
  Mutation: MutationResolvers<ContextType>;
};